カメラ角度制限実装 戦車オブジェクト追加

+++ UnlimitedGame/Assets/Scripts/Managers/Confug.cs
@@ -65,7 +65,7 @@ public class Confug : MonoBehaviour
     void Update()
     {
         _menu.SetActive(_confugActive);
-        _mouseMove = _mouseSlider.value;
+        _mouseMove = (int)_mouseSlider.value;
         _sliderValume.text = _mouseMove.ToString();
         if (Input.GetKeyDown(KeyCode.Escape))
         {
